+++ internal/component/metrics/dud_type.go
@@ -17,6 +17,15 @@ func (d DudStat) Timing(delta int64) {}
 // Set does nothing.
 func (d DudStat) Set(value int64) {}
 
+// SetFloat64 does nothing
+func (d DudStat) SetFloat64(value float64) {}
+
+// IncrFloat64 does nothing
+func (d DudStat) IncrFloat64(count float64) {}
+
+// DecrFloat64 does nothing
+func (d DudStat) DecrFloat64(count float64) {}
+
 //------------------------------------------------------------------------------
 
 var _ Type = DudType{}