+++ internal/impl/aws/metrics_cloudwatch.go
@@ -119,6 +119,18 @@ type cloudWatchStat struct {
 	dimensions []*cloudwatch.Dimension
 }
 
+func (c *cloudWatchStat) SetFloat64(value float64) {
+	c.Set(int64(value))
+}
+
+func (c *cloudWatchStat) IncrFloat64(count float64) {
+	c.Incr(int64(count))
+}
+
+func (c *cloudWatchStat) DecrFloat64(count float64) {
+	c.Decr(int64(count))
+}
+
 // Trims a map of datum values to a ceiling. The primary goal here is to be fast
 // and efficient rather than accurately preserving the most common values.
 func trimValuesMap(m map[int64]int64) {
@@ -184,7 +196,7 @@ func (c *cloudWatchStat) addValue(v int64) {
 	c.root.datumLock.Unlock()
 }
 
-// Incr increments a metric by an amount.
+// Incr increments a metric by an int64 amount.
 func (c *cloudWatchStat) Incr(count int64) {
 	c.addValue(count)
 }